fileHandler comments

+++ server/fileHandler.py
@@ -4,7 +4,13 @@ import os
 class FileHandler():
     files = {}
 
+    # Creates new file and stores it.
+    # If File with same name exists, overwrites it.
+    # @port {string} Identifier of websocket
+    # @format {string} Must be either "net" or "trips". Labels file type
     def handleNewFile(self, port, format):
+        if format != "net" or format != "trips":
+            return
 
         try:
             os.remove(f"../sumo/upload{port}.{format}.xml")
@@ -16,17 +22,26 @@ class FileHandler():
 
             self.files[port][format] = open(f"../sumo/upload{port}.{format}.xml", 'ab')
 
+    # Writes data into the file.
+    # @port {string} Identifier of websocket
+    # @message {bites} Data that are beeing written to the file
     def appendToFile(self, port, message):
         if "net" in self.files[port] and self.files[port]["net"].closed == False:
             self.files[port]["net"].write(message)
             return
 
-        if "trip" in self.files[port] and self.files[port]["trip"].closed == False:
-            self.files[port]["trip"].write(message)
-            
-    def closeFile(self, port, format):
-        self.files[port][format].close()
+        if "trips" in self.files[port] and self.files[port]["trips"].closed == False:
+            self.files[port]["trips"].write(message)
 
+    # Closes file. It is important for appendToFile function, because closed files cannot be written to.
+    # @port {string} Identifier of websocket
+    # @format {string} Must be either "net" or "trips". Labels file type
+    def closeFile(self, port, format):
+        if port in self.files and format in self.files[port]:
+            self.files[port][format].close()
+
+    # Removes uploaded files from directory and their refference in this class
+    # @port {string} Identifier of websocket
     def removeFile(self, port):
         try:
             os.remove(f"../sumo/upload{port}.trips.xml")
